Open collapsible when a validation error occurs inside
- Rename CollapsibleCourseRunFields -> CollapsibleCourseRuns
- Add CollapsibleCourseRun component to hold one single run
- Add isInvalid handlers for fields that will open the collapsible

Not all fields currently report isInvalid - that's a work in progress.

DISCO-882
