+++ src/sass/App.scss
@@ -8,7 +8,7 @@ $fa-font-path: "~font-awesome/fonts";
 
 // Components
 @import '../components/EditCoursePage/EditCourseForm';
-@import '../components/EditCoursePage/CollapsibleCourseRunFields';
+@import '../components/EditCoursePage/CollapsibleCourseRun';
 @import '../components/Header/Header';
 @import '../components/StaffList/StaffList';
 @import '../components/StafferPage/StafferForm';
